pyproject.toml
[tool.poetry]
name = "petshop"
version = "0.1.0"
description = "A Python package browser and search engine"
authors = [ "Frank Ploss <mail@fqxp.de>" ]
license = "GPL-3.0-only"
readme = "README.md"
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Framework :: FastAPI",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-bigquery = "^3.26.0"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.22"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
fastapi = {extras = ["standard"], version = "^0.115.5"}
poetry-plugin-dotenv = "^2.5.1"
uvicorn = "^0.32.1"
python-dateutil = "^2.9.0.post0"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
google-api-python-client-stubs = "^1.28.0"
alembic = "^1.14.0"
poetry-plugin-dotenv = "^2.5.1"
pytest = "^8.3.3"

[tool.poetry.scripts]
petshop-import = "petshop.importer.cli:cli"
petshop-server = "petshop.web.main:main"

[tool.pyright]
reportUnusedExpression = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
